Server/client/repair information

+++ CarShowroom/Models/Maintenances/MaintenanceStatisticsGridModel.cs
@@ -4,7 +4,7 @@
     {
         public string Vehicle { get; set; }
 
-        public float TotalHours { get; set; }
+        public double TotalHours { get; set; }
 
         public decimal TotalCost { get; set; }
     }
